rename to ConsoleTable
